rangeに入れる数値をintにキャスト

+++ ros2_ugokuita_ws/src/collision_lidar_pkg/collision_lidar_pkg/collision_lidar_subscriber.py
@@ -21,7 +21,7 @@ class ControllerSubscriber(Node):
 
     def listener_callback(self, msg):
         half_degrees = 50 # degrees
-        half_range = 502/2 * half_degrees / 360
+        half_range = int(502/2 * half_degrees / 360)
         collision_distance_threshold = 1.5
         for i in range(251 - half_range, 251 + half_range):
             value = msg.ranges[i]
